requirements.txt
# requirements.txt

streamlit
python-dotenv
requests
langchain
langchain-google-genai
langgraph
flask